\name{rainbow-package}
\alias{rainbow-package}
\alias{rainbow}
\docType{package}
\title{
Rainbow Plots, Bagplots and Boxplots for Functional Data
}
\description{
This package presents the rainbow plots, bagplots and boxplots for functional data. The latter two can also be used to identify outliers, which have either the lowest depth or the lowest density.
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <hanlin.shang@anu.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2008) "Bagplots, boxplots and outlier detection for functional data", in S. Dabo-Niang and F. Ferraty, eds, `Functional and Operatorial Statistics', Springer, Heidelberg, pp. 201-207.

R. J. Hyndman and H. L. Shang. (2010) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Journal of Computational and Graphical Statistics}, \bold{19}(1), 29-45.
}
\keyword{package}

